Move ase db to single function - handling both .db and .xyz files.
